docker/flask-app/app/requirements.txt
Flask
redis
cassandra-driver
